+++ documentation/RELEASE_NOTES.md
@@ -1,3 +1,11 @@
+# Release 1.15.2 (11/25/2025)
+- Fix bug to make sure mouse over data match plot
+
+# Release 1.15.1 (11/17/2025)
+- highlight mouse over points in plot
+- fix xaxes config setting
+- several fixes to support mouse over when xaxes is set
+
 # Release 1.15.0 (11/12/2025)
 - upgraded to pyqtgraph 0.13
 - refactored configuration code, implemented configuration for rest of striptool and scope app settings